refactor(tests): move renderCodeEditor utility to CodeEditor.testUtils for better organization
