+++ packages/code-editor/src/CodeEditor/CodeEditor.testUtils.tsx
@@ -7,7 +7,7 @@ import {
   CodeEditorProps,
   CodeEditorSelectors,
   CodeMirrorView,
-} from '..';
+} from '.';
 
 let editorView: CodeMirrorView;
